requirements.txt
Django==3.1.5
gunicorn==19.9.0
psycopg2==2.8.3
django-filter==2.4.0
djangorestframework==3.12.2
django-bootstrap4==2.3.1
redis==3.4.1

# keep until merged
# https://github.com/agoragames/leaderboard-python/pull/62
git+git://github.com/YuriHeupa/leaderboard-python.git@ccf3b284db35512950909aed8d1d3436557d3efb
